requirements.txt
httpunixsocketconnection==1.0.1
jupyterlab>3.0
pydantic-yaml==1.4.0
pyyaml==6.0.1
streamlit-autorefresh==1.0.1
streamlit-extras==0.5.5
streamlit==1.41.1
requests
